+++ docs/source/release_notes.rst
@@ -5,8 +5,27 @@
 Releases
 ========
 
+Version 1.0.0-beta2
+-------------------
+
+Notes
+  * Update recommended
+  * New modules
+
+    * ims_acb_gen
+  * Bug fixes
+  * Documentation updates
+  * Updated sample playbook
+
+Availability
+  * Galaxy
+  * GitHub
+
+Reference
+  * Supported by IBM z/OS core collection 1.0.0 or later
+
 Version 1.0.0-beta1
--------------
+-------------------
 
 Notes
   * Initial beta release of IBM z/OS IMS collection, referred to as ibm_zos_ims
@@ -19,6 +38,4 @@ Notes
 Availability
   * Galaxy
   * GitHub
-
-Reference